feat: update Profile page styles and improve resource card design
- Adjusted padding and margin in Profile header for better spacing
- Changed background color of resource cards to white
- Enhanced hover effect with increased box shadow
- Updated resource icon dimensions and added border radius
- Modified resource title font weight for better emphasis

feat: enhance Profile component with dynamic resource styling

- Updated resource data structure to include gradient backgrounds and accent colors
- Applied dynamic styles to resource cards based on data attributes
- Improved accessibility by adjusting icon brightness

feat: add TimeValidate utility for flexible time parsing

- Implemented TimeValidate class to parse various time formats
- Added regex patterns for year, month, day, and time validation
- Created TimeRange record to encapsulate start and end times

feat: implement ActionHistoryController for action history retrieval

- Added ActionHistoryController to handle action history requests
- Implemented filtering and sorting functionality for action history

fix: correct DataSensorController request handling

- Changed request body parameters to query parameters for filtering

feat: define ActionHistoryDto and DataSensorDto for data transfer

- Created ActionHistoryDto to encapsulate action history data
- Updated DataSensorDto to use Float for sensor values

feat: enhance ActionHistoryRepository with filtering capabilities

- Added custom query to ActionHistoryRepository for filtered action history retrieval

feat: implement DataSensorRepository for sensor data retrieval

- Added filtering capabilities to DataSensorRepository for sensor data

feat: update ActionHistoryService implementation

- Implemented filtering and sorting logic in ActionHistoryService

fix: correct DataSensorService implementation

- Fixed class name typo in DataSensorService implementation

fix: update DeviceControlServiceImpl for action history logging

- Moved action history logging to correct location in device control logic

chore: update application properties for Swagger UI configuration

- Added Swagger UI configuration to use openapi.yaml file

docs: add README.md for project overview and setup instructions

- Documented project features, technologies, architecture, and setup instructions

feat: add OpenAPI specification for API documentation

- Created openapi.yaml file to define API endpoints and data models

+++ backend/src/main/java/com/example/iot/domain/dto/DataSensorDto.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 public record DataSensorDto(
         SensorType sensorType,
-        Double value,
+        Float value,
         LocalDateTime recordAt
 ) {
 }
